backend/requirements.txt
# Core
django==5.2.11
djangorestframework==3.16.1
mongoengine==0.29.1
django-cors-headers==4.9.0
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
django-filter==25.2

# ASGI server (WebSocket + HTTP)
daphne==4.2.1
channels==4.3.2
channels-redis==4.3.0

# WSGI fallback / static files
gunicorn==21.2.0
whitenoise==6.7.0

# Async / Celery
celery==5.6.2
redis==7.2.1

# Auth
bcrypt==4.2.0
pyotp==2.9.0
google-auth==2.35.0

# Integrations
google-api-python-client==2.149.0
google-auth-oauthlib==1.2.1
stripe==14.4.0
requests==2.31.0

# AI/ML
anthropic==0.84.0
scikit-learn==1.7.2

# PDF (désactivé en prod — weasyprint requiert libpango/libcairo non dispo sur free tier)
# weasyprint==62.3

# Utils
pillow==10.2.0
python-dotenv==1.0.0
mongomock==4.3.0

# Prod
pymongo==4.10.1
